+++ Crypt.php
@@ -27,6 +27,8 @@ class Crypt
                 return APR1::hash($password);
             case PasswordFile::ALG_SHA1:
                 return self::sha1($password);
+            case PasswordFile::ALG_CRYPT:
+                return self::cryptHash($password);
             case PasswordFile::ALG_PLAIN:
                 return $password;
         }
@@ -51,6 +53,9 @@ class Crypt
         if ($hash === self::sha1($password)) {
             return true;
         }
+        if (self::cryptVerify($password, $hash)) {
+            return true;
+        }
         return false;
     }
 
@@ -63,5 +68,31 @@ class Crypt
     public static function sha1($password)
     {
         return '{SHA}'.base64_encode(sha1($password, true));
+    }
+
+    /**
+     * Verifies a hash of CRYPT algorithm
+     *
+     * @param string $password
+     * @param string $hash
+     * @return bool
+     */
+    public static function cryptVerify($password, $hash)
+    {
+        $salt = substr($hash, 0, 2);
+        return (crypt($password, $salt) === $hash);
+    }
+
+    /**
+     * Hash a password using CRYPT algorithm
+     *
+     * @param string $password
+     * @return string
+     */
+    public static function cryptHash($password)
+    {
+        $salt = substr(base64_encode(chr(mt_rand(0, 255))), 0, 2);
+        $salt = str_replace('+', '.', $salt);
+        return crypt($password, $salt);
     }
 }