+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### dev-master
+
+* #1: Crypt algorithm
+
 ### 0.1.0 (29.11.2015)
 
 * Actions: add, edit, delete, verify